refactor: add bundle module
